Remove the clock from the Node relying on the Network clock. When a Node receives a Message it uses the Network clock as a timestamp of the receiving time.
If the Network is ticking ahead of the Node, when the Node ticks
it will fast forward to the Network time, receiving all the pending messages.

Exemplifying:
The Network ticks 5 times and receives a message and the Node has not ticked yet
(assuming it is offline for example),the Network ticks once more, going to the tick 6,
when the Node ticks, it will receive all the messages in the Network (in this case one)
and in its LogEntry the message will have a receive time of tick 7.

Add Candidate promotion logic;