Work in progress
Use tiny types for the Address, differentiating between
Source and Destination.

This causes a small bug not allowing to retrieve messages from the
Network when the Host and Port are the same but the type of the Address
is different.

It was considered to override the Hashcode and Equals to consider only
the data and not the type but that would break the contract of the language.
A more appropritated solution will be to use the Host and Port as keys
for the messages on the Network instead of the Address itself, to be fixed on the next commit.