Network uses Host and Port to retrieve the messages.
Using the Address as a key to retrieve the messages do not work
anymore as we are using to specialised types to differentiate the
Addresses (Source and Destination). Even if the types have the same
values (host and port) they are still considered different objects.

Instead of breaking the language contract of the 'equals and hashcode'
we use the Host and Port as a combined key to retrieve the messages.