Candidate, Follower and Leader return the new 'Node' after processing the incoming messages
The next step is to move the logic in 'tick' to its
appropriate pattern matchin in 'process'. For example, move the 'shouldBecomeLeader'
method in Candidate to the VoteFromFollower match in 'process'.

This allowed us to write a version of 'tick' without side effects: tickWithoutSideEffects.
We still might want to remove the side effects completely out of the Nodes.