When processing a message we add the message to the Node, and then (#7)
process the message. All the messages to be sent (responses) as a
result of processing are added in the `sent` messages of the Node
instead of being returned from `process`.

Move `process` and `tick` to the Node.
Add `add` methods to the Nodes to easily add Received and Sent messages.
Add Message.toSent() and Message.toReceived(), likely to change in
future.