Add buffer to separate sent messages from the not sent yet.
After receiving messages a Node does it internal processing, gets promoted
or updates its internal state, at the end messages are also a product of the
processing. This PR adds a 'buffer' to know which messages were sent previously and
which ones are new, as a result of the current process.