Adds paramters for the AppendEntry
For the Follower to compute if the AppendEntry message should be
accepted we need the prevLogIndex, prevLogTerm etc params.

In the Leader we are assuming we must increment the commitLog as soon
as we append an entry to the log and we are calculating the previous
mentioned params. We must revisit this as the index of the log has to
start with 1.