Fix commitIndex calculation
We create an updateCommit index so it can use the current state of the log.

If the Leader Commit Index is higher than the Followe Commit Index
the Follower needs to update its Commit Index and if the Leader Commit
Index is smaller than the log size we update the Follower Index to the
Leader Index.

This happens when the leader and the follower have been exchanging
messages but not yet committed, due to a delayed 3rd node.