Fix Election Bug - Leader not being elected
When an Election starts, the Candidate increments its Term
And should send a Request For Votes with the new Term.
The other Node, when receiving the Request with a higher Term
should update its own Term and send a Vote Response with the
newly updated Term.

There was a bug in the implementation where the Request For Vote
was being generated before incrementing the Term. That way the Follower
would vote with its outdated Term and the Candidate would not accept the
Vote, due to its term to be lower.

-----------------------------------------------------------------------

In the Election another bug was fixed where the Follower pointed at
itself as Peer instead of pointing to the other Node.

This have not affected the test/system because a Vote Response is sent
to the Address that originated the Vote Request, the Peers' list of the
Node is not checked in this case.

There is no invariant to check that a Node can communicated with only
its Peers. It is assumed all Nodes in the Networkd are well intentioned.

-----------------------------------------------------------------------

The Election Timeout reset was removed after the Node being promoted to
Leader because it was done when the vote is accepted.