Fix Election cancellation
This is not a bug or wasn't caught by the tests because termStartedAt
is set on the promotion to Candidate, this attribute does not bellong
to the othe Node types. Instead of setting to a 'magical' value, it
is more appropriate to set to null to signal that value is not utilised
at the current Node 'state/type'.

Setting it to null helps the computations to break and not make wrong
calculation if the value is not updated while promoting the Node.

The refactoring to have different types to represent the Node states
will solve the problem.