Log has Entries and not Messages
Entries in the Log need to have only Content and Term.
Addresses are concepts only of the Messages, although the
Message still have the Term it was sent from.

Most of the Messages will not need the Content.
Only AppendEntries and it is in a form of a List of Entries.
The 'content' attribute was not removed yet to maintain to change
as little as possible of the previous implementations. As well a new
type (AppendEntries) was introduced for the same reason.