Leader send Heartbeats
When the Heartbeat timeout is reached, the Leader
Replicate its Log to all the Followers.

This commit implements only the timeout checks
and resets the Heartbeat timeout clock.

The Log replication mechanism is not yet implemented.
Altough the previous commit has made corrections in the
design of the log.

Readme: raft(4/9) - Broadcasting Messages