Cancel election timeout instead of reseting it
Some Node attributes should not be used depending
on its state. For example the election timeout
(termStartedAt) should not be present in a Leader.

Instead of reseting it (set to 9) it should be canceled
(set to null), this helps in case the attribute is used
in a state it does not belong to, avoiding silent failures
or wrong computations.

Later on there will be a refactor into a type hierarchy and
the design will solve this problem.

This commit fix a bug where the election timeout should not get
canceled in a Candidate during an election. Only the Candidated
gets promoted to leader the timeout must be canceled.

When a Candidate receives a request for votes and become a Follower
the timeout also must be canceled and not reset, as it does not
belong to the Follower type.

The commit also has a small refactory where the size of the cluster
is not in a global constant anymore, and passed through configuration
to test scenarios where the Candidate promotes or not depending on the
quorum.