Only Leader must have sentLength and ackLength
Ensures only the Leader has sentLength and ackLength.
This class attributes are set during a promotion (currently
happening in the VoteHandler).

Previously this values were set in the Node initialization, so
it was available for all states, this would allow the values to
be used even if the Node wasn't a Leader, causing silent failures.

The attributes now can be null avoiding silent failure. However,
the silent failures are still possible in case of a demotion,
when the values are not canceled.

A refactor will be done to have each state represented by its type
avoiding this problem.