Bug Fix: Check Node's role and timeouts after message processing
After processing a Message the Node can have changed its Role as
well as reset or canceled timeouts. The checks were being performed
on the outdated Node instance.

This was caught on an integration test where the Node was starting
an election due to communication timeout even tho it has received a
message reseting that clock, because the check was on the instance
that had received the message, not on the instance after the message
been processed.