Leader does not add itself to the ackedLength on its promotion
In Kleppmann's implementation the only place where the Leader
add itself to the ackedLength record is when it receives a
command from the client.

I thought the Leader had to include itself on the record on the
promotion because on the last part of the algorithm (when the leader
commits its own entries after confirmation from the Followers),
the quorum of Nodes with that entry must be achieved, counting
included the Leader.

Klepmmann is correct on adding the Leader to ackedLength only at
the client command because, when a new entry arrives it must be
counted on the leader. But in the case of an election, the recently
promoted Leader will already have a most up to date log.