Add Logs
Started adding logs.

Logs are in the past and future tense because, sometimes the change
in state is the last expression evaluated, in that case the log
has to be added before and use the future tense.

Also, no logic was changed, just expressions were assigned to variables
to make it easy to log the messages or to determine if the log should be
shown or not.

The pattern used is that general events should be logged as info
and each condition of the events should be logged as debbug.