Fix communicationTimeout bug
The Follower was starting a timeout when it had not received a message
for longer than the heartbeatTimeout.

The hearbeatTimeout should be used only in the Leader to signal when
the leader should send a heat beat to the followers.

If the follower uses the heartbeatTimeout to measure when to start an
election, the follower will be constantly starting elections with a
still valid leader.

The follower should start an election only when the electionTimeout has
been reached.