Fix Timeout tests
The commit b141f29588d0907e0ae6faa9666c22bcf3cff13d
broke the tests because of the change in the default
timeout configurations, and the tests used the default values.

However, the tests should have broken on the commit
2b40647ba811633b1f0a5a0a4c5f240c8bc101c3
because that was the commit changed the logic for the timeouts.